requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.1
streamlit==1.10.0
textrank4zh==0.3
torch==1.10.1
transformers==3.0.2
